backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
textblob==0.15.3
pandas==1.3.3
pyjwt==2.1.0
python-jose==3.3.0
python-dotenv==0.19.0
